Add files via upload
Note that these are just prototypes and that this is not yet finished. At the moment this only does partial coordinate conversion between cartesian and Toroidal coordinates, but when finished it will be capable of doing that in both directions, and of calculating derivatives/integrals as mentioned in the README file.