Add files via upload
Reinforcement learning notebook used on IonQ Harmony for testing out whether this can run on an Ion Trap Quantum Computer or not (i.e. NOT simulated).

Note that this may need debugging, as it produced more than 1 attempt at the job when last tested, although it did partially train the network as shown in the .doc file here.