/numpy-toroidal-coordinates/
Cartesian to Toroidal converter, used for both complex and real valued numerical arrays. 

Note that this can work with both, IF the imaginary component is set to 0 then it will work the same as if using real variables.

Note also that this is only a prototype, and as the author I ask for any use of this to involve giving me credit for my work. Other than that I have no problem with it being modified, distributed, or used in any appropriate context.

Thomas N Toseland.

P.S. My name must appear on any derived or otherwise distributed versions of this.