Add files via upload
The results from my attempt to run a reinforcement learning algorithm on quantum computing are shown here, but please note that I did NOT create the original algorithm, I just modified an existing one to see what would happen, if I used double-qubit rotation gates and complex variables on a real quantum computer instead of a simulated one. Namely, IonQ's Harmony which is an Ion-trap machine with 11 qubits at the time of testing.