Add files via upload
Ultralytics neural network source code, used to modify RTDETR and YOLO to work with quantum machine learning algorithms.

This is the only part of it that was modified, the rest is exactly the same as the ultralytics 8.3.37 package, so if you download this and then replace the nn folder inside the ultralytics package of that version, it will work with this as long as you have all the required dependencies (e.g. PyTorch, Pennylane, Pennylane lightning plugin, pennylane lightning tensor plugin), please see the readme file for details.