/results/lvis
These are results from the LVIS prototype mentioned earlier with the tensor networks in, the validation results will be added shortly and note that this was done only on a single RTX 3090, not on a big server farm. this model can also be used for highly dense scenes, and it has both a segmentation and a oriented bounding boxes head. It is based on YOLOE but it contains quantum tensor networks, complex valued loss functions, including contour integration around any singularities.

This was done using MERA and TTN Tensor networks, and with complex valued loss functions.